Fixed standalone single quotes getting matched.
Also fixes the issue where tags were not being colored properly when in a YAML list.
Example:
- <color[#ff69b8]>Hello
The color tag will be colored improperly and will stop coloring past the "]". It would also be colored as a command, which we don't really want.
The standalone single quotes would color the rest of the line if it found. Example:
Breadcrumb's Awesome Script
Because there is no second single quote, the whole rest of the line following the single quote will be colored. This is fixed with this commit.