Fixed coloring key until last colon.
Will now only match the first colon.
If something like a URL was put into a key like for a data script, the key color would color everything until the last colon, which could be the colon in "https:" or a port, ":8080". This is not intended. This commit fixes that issue, only coloring up to the first colon which is the key.
key (stop coloring here) : value