Keys will now match even with hyphens in them.
Previously, the `[^-#\n]` part of the regex would not match anything that had a `-`, `#`, or `\n`. The former would prevent matching of a command. But, if you had a hyphen in a key, then it would not match the entire thing because it saw a hyphen. Now, it looks for `- ` (hyphen WITH a space) to not match with a command.