Fix a few highlighting issues with tags.
If you used a tag in a fallback like: "<context.args.first||<[my_def]>>", the last ">" would not be colored. Now it should be colored as tags can now parse tags inside tags!

Doing something like "<&[warning]>" would cause the "[warning]" to not be colored like a def-bracket. Now it recognizes "&" as a valid tag base!